Capstone Project/Resources/requirements.txt
fastapi
uvicorn[standard]
neo4j
pandas
requests
python-dotenv
pydantic
